feat(metrics): Metrics extraction config, custom tags
