+++ relay-server/src/metrics_extraction/mod.rs
@@ -1,0 +1,1 @@
+pub mod transactions;
